uhhhhh balance

+++ include/ace/coroutines/promise.h
@@ -95,14 +95,14 @@ namespace ace::coroutines {
         or std::same_as<decltype(modeT::action()), std::suspend_always>);
 
     struct average_quants {
-        static constexpr long int window_size = 4;
-        alignas(8) long int _total_sum {};
+        static constexpr int window_size = 4;
+        alignas(8) int _total_sum {};
         alignas(8) uint64_t _curr_member = 0;
-        std::array<uint32_t, window_size> _members {};
+        std::array<int, window_size> _members {};
 
-        [[nodiscard]] long int value() const { return _total_sum / window_size; }
+        [[nodiscard]] int value() const { return _total_sum / window_size; }
 
-        [[nodiscard]] long int add(const long int& new_one) {
+        [[nodiscard]] int add(const int& new_one) {
             _total_sum = _total_sum + new_one - _members[_curr_member % window_size];
             _members[_curr_member % window_size] = new_one;
             ++_curr_member;
